Split correctly the rnn tests
